+++ src/emoji_data/utils.py
@@ -1,5 +1,5 @@
 import os
-from typing import BinaryIO, Iterator
+from typing import BinaryIO, Iterator, Tuple
 
 from pkg_resources import Requirement, resource_stream
 
@@ -12,51 +12,29 @@ TEST_DATAFILE_STREAM = resource_stream(
 )
 
 
-def read_data_file_stream_iterable(handle: BinaryIO) -> Iterator[str]:
+def read_data_file_iterable(handle: BinaryIO) -> Iterator[Tuple[str, str]]:
     for line_bytes in handle:
-        line_text = ''
+        content = ''
+        comment = ''
         line_bytes = line_bytes.strip()
         if line_bytes:
             if line_bytes[0] not in b'#;':
-                line_text = line_bytes.decode('utf-8').split('#', 1)[0].strip()
-        yield line_text
+                parts = [s.strip() for s in line_bytes.decode('utf-8').split('#', 1)]
+                content = parts[0]
+                if len(parts):
+                    comment = parts[1]
+        yield content, comment
 
 
 def reload_test_data():
     result = []
     TEST_DATAFILE_STREAM.seek(0)
-    for line in read_data_file_stream_iterable(TEST_DATAFILE_STREAM):
-        if not line:
+    for content, _ in read_data_file_iterable(TEST_DATAFILE_STREAM):
+        if not content:
             continue
-        line = line.split('#', 1)[0].strip()
-        code_points, qualified_type = (part.strip() for part in line.split(';', 1))
+        code_points, qualified_type = (part.strip() for part in content.split(';', 1))
         result.append((
             [int(s, 16) for s in code_points.split()],
             qualified_type
         ))
     return result
-
-
-class BaseDictContainer(type):
-    def __new__(mcs, name, bases, attrs):  # pylint:disable=bad-mcs-classmethod-argument
-        mcs._data = dict()
-        return super().__new__(mcs, name, bases, attrs)
-
-    def __setitem__(self, key, value):  # pylint: disable=C0203
-        self._data[key] = value
-
-    def __delitem__(self, key):  # pylint: disable=C0203
-        del self._data[key]
-
-    def __getitem__(self, key):  # pylint: disable=C0203
-        return self._data[key]
-
-    def __contains__(self, key):  # pylint: disable=C0203
-        return key in self._data
-
-    def __iter__(self):  # pylint: disable=bad-mcs-method-argument
-        for k, v in self._data.items():  # pylint: disable=invalid-name
-            yield k, v
-
-    def __len__(self):  # pylint: disable=C0203
-        return len(self._data)